Promote TextureSet to its own class
Separate out several static functions into TextureSet
Rename SetEndsTextureProperties to more generic SetTextureProperties
